pyproject.toml
[project]
name = "opdapt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.31",
]
